5_custom_framework/1_mcp/pyproject.toml
[project]
name = "mcp-tools-server"
version = "0.1.0"
description = "MCP server providing tools for AI agents"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.10.1",
    "starlette>=0.27.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
    "tavily-python>=0.5.0",
    "python-dotenv>=1.0.0",
]
